old_common -> plotting, preprocessing
